Fixed a bug with deleting messages while gambling. Updated the output for lotto winners.

+++ uboot/dclient/bot.py
@@ -40,7 +40,7 @@ class DestructableView():
 
     async def remove(self) -> None:
         try:
-            if len(self.msg.content) == 0:
+            if len(self.msg.content) == 0 and len(self.msg.embeds) == 0:
                 return await self.msg.delete()
             self.msg = await self.msg.edit(view=None)
         except BaseException:
